syncing back to UI
