+++ src/contexts/UI/index.tsx
@@ -1,9 +1,17 @@
 // Copyright 2022 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import BN from 'bn.js';
 import { ServiceList, SideMenuStickyThreshold } from 'consts';
+import { ImportedAccount } from 'contexts/Connect/types';
+import { useActivePools } from 'contexts/Pools/ActivePools';
 import React, { useEffect, useRef, useState } from 'react';
 import { localStorageOrDefault, setStateWithRef } from 'Utils';
+import { useApi } from '../Api';
+import { useBalances } from '../Balances';
+import { useConnect } from '../Connect';
+import { useNetworkMetrics } from '../Network';
+import { useStaking } from '../Staking';
 import * as defaults from './defaults';
 import { UIContextInterface } from './types';
 
@@ -14,6 +22,22 @@ export const UIContext = React.createContext<UIContextInterface>(
 export const useUi = () => React.useContext(UIContext);
 
 export const UIProvider = ({ children }: { children: React.ReactNode }) => {
+  const { isReady } = useApi();
+  const { accounts: connectAccounts } = useConnect();
+  const { staking, eraStakers } = useStaking();
+  const { metrics } = useNetworkMetrics();
+  const { accounts } = useBalances();
+  const { synced: activePoolsSynced } = useActivePools();
+
+  // set whether the network has been synced.
+  const [networkSyncing, setNetworkSyncing] = useState(false);
+
+  // set whether pools are being synced.
+  const [poolsSyncing, setPoolsSyncing] = useState(false);
+
+  // set whether app is syncing.ncludes workers (active nominations).
+  const [isSyncing, setIsSyncing] = useState(false);
+
   // get initial services
   const getAvailableServices = () => {
     // get services config from local storage
@@ -81,6 +105,59 @@ export const UIProvider = ({ children }: { children: React.ReactNode }) => {
     resizeCallback();
   }, [userSideMenuMinimised]);
 
+  // app syncing updates
+  useEffect(() => {
+    let _syncing = false;
+    let _networkSyncing = false;
+    let _poolsSyncing = false;
+
+    if (!isReady) {
+      _syncing = true;
+      _networkSyncing = true;
+      _poolsSyncing = true;
+    }
+    // staking metrics have synced
+    if (staking.lastReward === new BN(0)) {
+      _syncing = true;
+      _networkSyncing = true;
+      _poolsSyncing = true;
+    }
+
+    // era has synced from Network
+    if (metrics.activeEra.index === 0) {
+      _syncing = true;
+      _networkSyncing = true;
+      _poolsSyncing = true;
+    }
+
+    // all extension accounts have been synced
+    const extensionAccounts = connectAccounts.filter(
+      (a: ImportedAccount) => a.source !== 'external'
+    );
+    if (accounts.length < extensionAccounts.length) {
+      _syncing = true;
+      _networkSyncing = true;
+      _poolsSyncing = true;
+    }
+
+    setNetworkSyncing(_networkSyncing);
+
+    // active pools have been synced
+    if (activePoolsSynced !== 'synced') {
+      _syncing = true;
+      _poolsSyncing = true;
+    }
+
+    setPoolsSyncing(_poolsSyncing);
+
+    // eraStakers total active nominators has synced
+    if (!eraStakers.totalActiveNominators) {
+      _syncing = true;
+    }
+
+    setIsSyncing(_syncing);
+  }, [isReady, staking, metrics, accounts, eraStakers, activePoolsSynced]);
+
   const setSideMenu = (v: number) => {
     setSideMenuOpen(v);
   };
@@ -123,6 +200,9 @@ export const UIProvider = ({ children }: { children: React.ReactNode }) => {
         userSideMenuMinimised: userSideMenuMinimisedRef.current,
         sideMenuMinimised,
         services: servicesRef.current,
+        isSyncing,
+        networkSyncing,
+        poolsSyncing,
         containerRefs,
       }}
     >
