+++ src/library/Form/NominateStatusBar/index.tsx
@@ -5,8 +5,8 @@ import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import { faFlag } from '@fortawesome/free-regular-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useApi } from 'contexts/Api';
-import { useSetup } from 'contexts/Setup';
 import { useStaking } from 'contexts/Staking';
+import { useUi } from 'contexts/UI';
 import { OpenHelpIcon } from 'library/OpenHelpIcon';
 import { useTranslation } from 'react-i18next';
 import { planckBnToUnit, toFixedIfNecessary } from 'Utils';
@@ -15,7 +15,7 @@ import { Wrapper } from './Wrapper';
 
 export const NominateStatusBar = ({ value }: NominateStatusBarProps) => {
   const { staking, eraStakers } = useStaking();
-  const { isSyncing } = useSetup();
+  const { isSyncing } = useUi();
   const { unit, units } = useApi().network;
   const { minNominatorBond } = staking;
   const { minActiveBond } = eraStakers;
