+++ src/library/StatusLabel/index.tsx
@@ -4,7 +4,6 @@
 import { faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { usePoolMemberships } from 'contexts/Pools/PoolMemberships';
-import { useSetup } from 'contexts/Setup';
 import { useStaking } from 'contexts/Staking';
 import { useUi } from 'contexts/UI';
 import OpenHelpIcon from 'library/OpenHelpIcon';
@@ -14,8 +13,7 @@ import { Wrapper } from './Wrapper';
 export const StatusLabel = (props: StatusLabelProps) => {
   const status = props.status ?? 'sync_or_setup';
 
-  const { isSyncing } = useSetup();
-  const { services } = useUi();
+  const { isSyncing, services } = useUi();
   const { inSetup } = useStaking();
   const { membership } = usePoolMemberships();
