+++ src/pages/Nominate/Active/ControllerNotImported.tsx
@@ -5,9 +5,9 @@ import { ButtonPrimary } from '@rossbulat/polkadot-dashboard-ui';
 import { useBalances } from 'contexts/Balances';
 import { useConnect } from 'contexts/Connect';
 import { useModal } from 'contexts/Modal';
-import { useSetup } from 'contexts/Setup';
 import { useStaking } from 'contexts/Staking';
 import { useTheme } from 'contexts/Themes';
+import { useUi } from 'contexts/UI';
 import { CardHeaderWrapper, CardWrapper } from 'library/Graphs/Wrappers';
 import { useTranslation } from 'react-i18next';
 import { defaultThemes } from 'theme/default';
@@ -15,7 +15,7 @@ import { PageRowWrapper } from 'Wrappers';
 
 export const ControllerNotImported = () => {
   const { openModalWith } = useModal();
-  const { isSyncing } = useSetup();
+  const { isSyncing } = useUi();
   const { mode } = useTheme();
   const { getControllerNotImported } = useStaking();
   const { activeAccount, isReadOnlyAccount } = useConnect();
