+++ src/pages/Nominate/Active/ManageBond.tsx
@@ -8,9 +8,9 @@ import { useApi } from 'contexts/Api';
 import { useBalances } from 'contexts/Balances';
 import { useConnect } from 'contexts/Connect';
 import { useModal } from 'contexts/Modal';
-import { useSetup } from 'contexts/Setup';
 import { useStaking } from 'contexts/Staking';
 import { useTransferOptions } from 'contexts/TransferOptions';
+import { useUi } from 'contexts/UI';
 import BondedGraph from 'library/Graphs/Bonded';
 import { CardHeaderWrapper } from 'library/Graphs/Wrappers';
 import useUnstaking from 'library/Hooks/useUnstaking';
@@ -27,7 +27,7 @@ export const ManageBond = () => {
   const { getLedgerForStash } = useBalances();
   const { getTransferOptions } = useTransferOptions();
   const { inSetup } = useStaking();
-  const { isSyncing } = useSetup();
+  const { isSyncing } = useUi();
   const ledger = getLedgerForStash(activeAccount);
   const { isFastUnstaking } = useUnstaking();
   const { active }: { active: BN } = ledger;
