+++ src/pages/Nominate/Active/UnstakePrompts.tsx
@@ -6,9 +6,9 @@ import { ButtonPrimary } from '@rossbulat/polkadot-dashboard-ui';
 import { useApi } from 'contexts/Api';
 import { useConnect } from 'contexts/Connect';
 import { useModal } from 'contexts/Modal';
-import { useSetup } from 'contexts/Setup';
 import { useTheme } from 'contexts/Themes';
 import { useTransferOptions } from 'contexts/TransferOptions';
+import { useUi } from 'contexts/UI';
 import { CardWrapper } from 'library/Graphs/Wrappers';
 import useUnstaking from 'library/Hooks/useUnstaking';
 import { useTranslation } from 'react-i18next';
@@ -20,7 +20,7 @@ export const UnstakePrompts = () => {
   const { activeAccount } = useConnect();
   const { mode } = useTheme();
   const { openModalWith } = useModal();
-  const { networkSyncing } = useSetup();
+  const { networkSyncing } = useUi();
   const { isFastUnstaking, isUnstaking, getFastUnstakeText } = useUnstaking();
   const { getTransferOptions } = useTransferOptions();
   const { active, totalUnlockChuncks, totalUnlocked, totalUnlocking } =
