+++ src/pages/Overview/Tips/index.tsx
@@ -14,9 +14,9 @@ import { useConnect } from 'contexts/Connect';
 import { useModal } from 'contexts/Modal';
 import { useActivePools } from 'contexts/Pools/ActivePools';
 import { usePoolMemberships } from 'contexts/Pools/PoolMemberships';
-import { useSetup } from 'contexts/Setup';
 import { useStaking } from 'contexts/Staking';
 import { useTransferOptions } from 'contexts/TransferOptions';
+import { useUi } from 'contexts/UI';
 import { CardHeaderWrapper, CardWrapper } from 'library/Graphs/Wrappers';
 import useFillVariables from 'library/Hooks/useFillVariables';
 import { OpenHelpIcon } from 'library/OpenHelpIcon';
@@ -32,7 +32,7 @@ import { PageToggleWrapper } from './Wrappers';
 export const Tips = () => {
   const { network } = useApi();
   const { activeAccount } = useConnect();
-  const { networkSyncing } = useSetup();
+  const { networkSyncing } = useUi();
   const { openModalWith } = useModal();
   const { fillVariables } = useFillVariables();
   const { membership } = usePoolMemberships();
